backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
databricks-sql-connector==4.2.5
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.135.3
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
lz4==4.4.5
Mako==1.3.11
MarkupSafe==3.0.3
numpy==2.4.4
oauthlib==3.3.1
openpyxl==3.1.5
pandas==2.3.3
passlib==1.7.4
psycopg2-binary==2.9.11
pyarrow==23.0.1
pyasn1==0.6.3
pybreaker==1.4.1
pycparser==3.0
pydantic==2.13.0
pydantic-settings==2.13.1
pydantic_core==2.46.0
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.26
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.1
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
thrift==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.44.0
watchfiles==1.1.1
websockets==16.0
